feat: update service

+++ back-end/app/Services/RateService.php
@@ -2,10 +2,10 @@
 
 namespace App\Services;
 
+use App\Exceptions\RateException;
+use App\Http\Resources\GeneralResource;
 use App\Http\Resources\RateResource;
 use App\Models\Rate;
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 
 class RateService
@@ -15,48 +15,44 @@ class RateService
         try {
             $rate = Rate::select('idItens', DB::raw('COUNT(*) as count'))->groupBy('idItens')->where('idItens', $id)->get();
             return new RateResource($rate);
-        } catch (\Throwable $th) {
-            throw $th;
+        } catch (\Exception $e) {
+            throw new RateException($e->getMessage());
         }
     }
 
     public function showLikeUser(string $id)
     {
         try {
-            $user = Auth::user();
-            $idUser = $user->idUser;
-            $rate = Rate::where('idUser', $idUser)->where('idItens', $id)->get();
+            $rate = Rate::where('idUser', auth()->user()->idUser)->where('idItens', $id)->get();
             if ($rate->isNotEmpty())
-                return response()->json(['data' => 'true']);
-            return response()->json(['data' => 'false']);
-        } catch (\Throwable $th) {
-            throw $th;
+                return new GeneralResource(['data' => 'true']);
+            return new GeneralResource(['data' => 'false']);
+        } catch (\Exception $e) {
+            throw new RateException($e->getMessage());
         }
     }
     public function store(array $data)
     {
         try {
-            $user = Auth::user();
-            $data['idUser'] = $user->idUser;
-            $exist = Rate::where('idUser', $data['idUser'])->where('idItens', $data['idItens'])->first();
-            
+            $exist = Rate::where('idUser', auth()->user()->idUser)->where('idItens', $data['idItens'])->first();
+
             if ($exist) {
-                return response()->json(['message' => 'Já avaliou'], 400);
+                return new GeneralResource(['message' => 'Já avaliou']);
             }
 
-            Rate::create($data);
-            return response()->json(['message' => 'sucess'], 200);
-        } catch (\Throwable $th) {
-            throw $th;
+            auth()->user()->rates()->create($data);
+            return new GeneralResource(['message' => 'success']);
+        } catch (\Exception $e) {
+            throw new RateException($e->getMessage());
         }
     }
     public function destroy(string $id)
     {
         try {
-            Rate::where('idItens', $id)->delete();
-            return response()->json(['message' => 'sucess'], 200);
-        } catch (\Throwable $th) {
-            throw $th;
+            Rate::where('idUser', auth()->user()->idUser)->where('idItens', $id)->delete();
+            return new GeneralResource(['message' => 'success']);
+        } catch (\Exception $e) {
+            throw new RateException($e->getMessage());
         }
     }
 }
